[fix] アコーディオンのコンポーネント化と他不要箇所の修正
