+++ resources/js/Components/SideMenu.tsx
@@ -1,60 +1,26 @@
 import { useState } from 'react'
 import { Link } from '@inertiajs/inertia-react'
-import { routeList } from '@/data/route'
+import Accordion from '@/Components/Accordion'
+import routeList from '@/data/route'
 import ApplicationLogo from '@/Components/ApplicationLogo'
 
 const SideMenu = () => {
-    const [activeItem, setActiveItem] = useState<number[]>([])
-
-    const toggleItem = (index: number) => {
-        if (activeItem.includes(index)) {
-            setActiveItem((prevItems) =>
-                prevItems.filter((itemIndex) => itemIndex !== index)
-            )
-        } else {
-            setActiveItem((prevItems) => [...prevItems, index])
-        }
-    }
-
     return (
         <div className="relative lg:block hidden">
             <div className="bg-black text-slate-200 lg:w-48 h-screen absolute left-0">
                 <div className="flex justify-center mt-3">
                     <Link href="/" className="flex items-center">
                         <ApplicationLogo className="block h-9 w-auto fill-white" />
-                        <span className="ml-3">sidemnue</span>
+                        <span className="ml-3">My Routing</span>
                     </Link>
                 </div>
 
                 <div className="mt-10">
-                    {routeList.map((item, index) => (
-                        <div key={index} className="cursor-pointer">
-                            <div
-                                onClick={() => toggleItem(index)}
-                                className="pl-10 h-10 font-semibold flex items-center hover:bg-light-gray duration-300"
-                            >
-                                {item.title}
-                            </div>
-                            {item.itemLists && (
-                                <ul
-                                    className={`transition-all duration-500 ${
-                                        activeItem.includes(index)
-                                            ? 'max-h-96'
-                                            : 'max-h-0'
-                                    } overflow-hidden`}
-                                >
-                                    {item.itemLists.map((listItem) => (
-                                        <li
-                                            key={listItem}
-                                            className="pl-10 h-10 bg-gray hover:bg-light-gray flex items-center"
-                                        >
-                                            {listItem}
-                                        </li>
-                                    ))}
-                                </ul>
-                            )}
-                        </div>
-                    ))}
+                    <Accordion
+                        listItems={routeList}
+                        titleClsddName="pl-10 h-10 font-semibold flex items-center justify-start hover:bg-light-gray duration-300"
+                        itemListsClsddName="pl-10 h-10 bg-gray hover:bg-light-gray flex items-center"
+                    />
                 </div>
             </div>
         </div>
