+++ resources/js/types/ListItems.ts
@@ -1,0 +1,4 @@
+export default interface ListItems {
+    title: string
+    itemLists?: string[]
+}
